another 32 bit fix

+++ include/coverage-32.h
@@ -47,7 +47,7 @@ void simplify_trace(afl_state_t *afl, u8 *bytes) {
 
 inline void classify_counts(afl_forkserver_t *fsrv) {
 
-  u64 *mem = (u32 *)fsrv->trace_bits;
+  u32 *mem = (u32 *)fsrv->trace_bits;
   u32  i = (fsrv->map_size >> 2);
 
   while (i--) {
